ocr/requirements.txt
Pillow==8.4.0
pyperclip==1.8.2
pyscreenshot==3.0
requests==2.26.0
